Colored faces try 2
I think I'm close... the three.js editor does not show colors, but if
you debug the scene they are all there in the MeshFaceMaterial.  I'm
wondering if we need to call ComputeNormals or computeFaceNormals or
computeVertexNormals on the mesh after it's been created.  Might have to
try this in our own viewer, or i might just be messing something up  ...
harsh tokes today mr doob.